Initial setup for Ex3
This is the Choose Your Own Adventure exercise. This sets up a few types
for parsing a cyoa.json file. The story type is the entire file. Stories
have chapters, chapters have a Title, Paragraph (both strings) and then
an Option struct. The option type has a choices and then the
corresponding chapter name.

The main.go file currently will open the passed in file, decode, then
output the contents.

This is just POC json parsing.